requirements.txt
langchain==0.3.27
langchain-openai==0.3.32
langchain-postgres==0.0.15
langchain-community==0.3.29
langchain-text-splitters==0.3.11
pypdf==6.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
openai==1.106.1
